requirements.txt
Django==5.1.5
django-import-export==4.3.4
django-tinymce==4.1.0
weasyprint==63.1
shortuuid==1.0.13